third_party/blink/web_tests/platform/win/external/wpt/offscreen-canvas/text/2d.text.measure.actualBoundingBox-expected.txt

This is a testharness.js-based test.
FAIL Testing actualBoundingBox for OffscreenCanvas assert_equals: Math.abs(ctx.measureText('A').actualBoundingBoxLeft) === 0 (got 1[number], expected 0[number]) expected 0 but got 1
Harness: the test ran to completion.
